Fix "not loaded" error when specifying a custom DLL
This happens because on Python 3.8+ the DLL load path is modified to include
only system directories; not even the directory in which the DLL is located.
See: https://docs.python.org/3/whatsnew/3.8.html#changes-in-the-python-api

Without these changes, it would be impossible to load a DLL file that has
dependencies on other DLL files like that of Mesa, even when the full path is
specified. In this change, the DLL and its dependencies are resolved as
done in ctypes, first system directories and then the directories in which
the DLL file specified exists.

Signed-off-by: Naveen M K <naveen521kk@gmail.com>